training/m3/requirements_m3.txt
pandas
numpy
mlflow
pyyaml
prophet
scipy
